pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nemo-guardrails-poc"
version = "0.1.0"
description = "Proof-of-concept: NVIDIA NeMo Guardrails applied to a vulnerable LLM agent"
requires-python = ">=3.10"
dependencies = [
    "nemoguardrails>=0.9.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
agent       = "agent.main:main"
agent-guarded = "guardrails.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["agent", "guardrails"]
